requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-crontab==0.7.1
django-db-logger==0.1.13
django-debug-toolbar==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
pdfkit==1.0.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
